fix prognosis result first month balance
Before calculating the value for the output table, it will now be checked whether the salary for the previous month has been posted.

If it is posted, insert 0 into the value for the first month in the table. If it is not posted, assign the average salary to the first value.

The value is checked only for the current month.